+++ changelog.md
@@ -3,6 +3,7 @@
 ## 1.16.2
 * `NEW` diagnostic: `duplicate-set-field`
 * `NEW` diagnostic: `no-implicit-any`, disabled by default
+* `CHG` completion: improve field and table
 * `FIX` diagnositc: may open file with wrong uri case
 * `FIX` [#406](https://github.com/sumneko/lua-language-server/issues/406)